+++ src/site/pages/pages.11tydata.js
@@ -2,6 +2,6 @@ module.exports = function () {
   return {
     contentType: 'page',
     layout: 'page',
-    permalink: './{{ slug }}/index.html'
+    permalink: './site/{{ slug }}/index.html'
   }
 }