+++ src/site/projects/projects.11tydata.js
@@ -2,6 +2,6 @@ module.exports = function () {
   return {
     contentType: 'project',
     layout: 'project',
-    permalink: './portfolio/{{ slug }}/index.html'
+    permalink: './site/portfolio/{{ slug }}/index.html'
   }
 }