+++ _scripts/eleventyDebug.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+echo "Building $PLATFORM to build/$PLATFORM and debugging"
+DEBUG=Eleventy* npx eleventy --output=build/$PLATFORM --dryrun