+++ _scripts/eleventyServe.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+echo "Building $PLATFORM to build/$PLATFORM and serving"
+npx eleventy --output=build/$PLATFORM --serve