+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add npm scripts (`package.json`)
 - Add assets, data, includes, layouts, content (`/src/*`)
 - Add environment variables (`.env`)
+- Add create:content script
 
 ### Changed
 - Move netlify and now config files to `/_deploy/*`
@@ -29,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update GitHub Issue and Pull Request templates
 - Update lighthouse script and move into `/_scripts/lighthouse.sh`
 - Move lighthouse logs into `/_logs/lighthouse`
+- Update Netlify config (`/netlify.toml`)
+- Update page markup
 
 ### Fixed
 - Fix API post content