+++ CHANGELOG.md
@@ -22,10 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add npm scripts (`package.json`)
 - Add assets, data, includes, layouts, content (`/src/*`)
 - Add environment variables (`.env`)
-- Add create:content script
+- Add scripts
 
 ### Changed
-- Move netlify and now config files to `/_deploy/*`
 - Move assets and content into platform directories (`app`, `cms`, `email`, `site`)
 - Update GitHub Issue and Pull Request templates
 - Update lighthouse script and move into `/_scripts/lighthouse.sh`